refactor: acceptance 테스트 패키지 생성
