+++ src/test/java/com/server/crews/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package com.server.crews.environ.acceptance;
+package com.server.crews.acceptance;
 
 import static com.server.crews.fixture.RecruitmentFixture.RECRUITMENT_SAVE_REQUEST;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
